slist: rewrite slist_delete()
The previous implementation had a serious flaw: the loop used **start as
iteration variable and changed it, so before each return we would have
to make sure it was set exactly right. Now we use a local iteration
variable and only touch **start and **last if necessary.

Additionally the memory of the deleted element is freed now. This was
not done before, but after removing the item from the list, there was no
way to ever release this memory anymore: memory leak!

The new implementation was tested with unit tests in a personal branch.

Fixes #23.

Tested-by: Alexander Dahl <post@lespocky.de>