session: Added API function to free session internals
This fixes a possible memory leak when working with sessions (especially
with some non-mainline fcgi patches). Quoting from the ticket:

The global variable `sess_fname` is a pointer to dynamically allocated
memory which happens either in `sess_generate_id()` or in
`cgi_session_start()`. However this memory is never freed.

To not expose the variable globally or even just to `cgi.c` where e.g.
`cgi_end()` could access it, a new API function is introduced. We leave
bumping the API version up to a follow up commit.

Fixes: #25
Signed-off-by: Alexander Dahl <post@lespocky.de>