Use check instead of assert
assert() has the problem it is completely ignored in non Debug builds
and so those checks were not executed there. This lead to failing tests
when setting CMAKE_BUILD_TYPE to Release and RelWithDebInfo, probably
because of compiler optimizations. Replacing the assert() calls with the
check() macro we already use for the test itself, solves this problem.