Make BUILD_SHARED_LIBS optional
PR #39 changed the behaviour to support building a shared or static
library based on the global CMake variable BUILD_SHARED_LIBS, but the
build system had to set it from outside. Now it's an option and the user
can easily set it in the CMake GUI, while setting from outside is of
course still possible. Additionally this allows us to change the default
from building a static library to building a shared library like it was
before PR #39.