Mark function md5() as deprecated and compile md5.c again
Quote from issue #15:

> libcgi contains an implementation of the md5 hash function, however
> its not used inside the library itself and the status regarding
> license or potential security issues (despite the weekness of md5
> itself) is not clear. If a user needs md5 he or she could get it from
> a specialized library.

Although md5 is deprecated (#15) we can not simply remove the code from
the API. Do that later.

The HAVE_MD5 macro is useless, that was a relict from the old autotools
build and actually meant »build with md5«, so we can simply drop the
build time generated file and that macro.

Fixes: 0643151cf9ee8e09f4a613256bada8cde24b75c4
Signed-off-by: Alexander Dahl <post@lespocky.de>